+++ src/settings/index.ts
@@ -9,3 +9,19 @@ export {
   loadSettings,
   serverSettingsShape,
 } from "@/settings/base";
+export {
+  authSettingsShape,
+  emailSettingsShape,
+  envBoolean,
+  envList,
+  jwtSettingsShape,
+  logSettingsShape,
+  minioSettingsShape,
+  rabbitmqSettingsShape,
+  redisSettingsShape,
+  sessionSettingsShape,
+  tokenSettingsShape,
+  uploadSettingsShape,
+  webPushSettingsShape,
+  webSocketSettingsShape,
+} from "@/settings/mixins";